refactor(dftu_pw): split into dftu_tools_pw.{h,cpp} and dftu_cal_occ_pw.cpp
Step 4d-prerequisite of the dftu_pw decoupling plan. Enforce the
"header-implementation pairing" rule: the free-function declarations in
dftu_pw.h had their implementations stuffed at the bottom of dftu_pw.cpp
(alongside Plus_U_Base member functions), which is not a proper .h/.cpp
pair.

File structure after this commit:
- dftu_tools_pw.h      (renamed from dftu_pw.h)
    - declare 5 free functions in namespace dftu_pw
    - include guard DFTU_TOOLS_PW_H
- dftu_tools_pw.cpp    (new file)
    - implement the 5 free functions (pauli_to_spin_basis,
      compute_vu_spinor, compute_vu_scalar, accumulate_occ_spinor,
      accumulate_occ_scalar); depends only on <complex> and matrix.h
- dftu_cal_occ_pw.cpp  (renamed from dftu_pw.cpp via git mv)
    - Plus_U_Base member functions only (cal_occ_pw, accumulate_occ_one_k,
      reduce_occ_mat, sync_occ_to_uom, compute_eff_pot_and_energy)
    - includes dftu_tools_pw.h to call the free functions

namespace stays dftu_pw (not renamed) to avoid churning all call sites;
only filenames change.

Files changed:
- source/source_pw/module_pwdft/dftu_tools_pw.h (git mv from dftu_pw.h)
    - update include guard to DFTU_TOOLS_PW_H
- source/source_pw/module_pwdft/dftu_tools_pw.cpp (new)
    - body moved verbatim from the former dftu_pw.cpp namespace block
- source/source_pw/module_pwdft/dftu_cal_occ_pw.cpp (git mv from dftu_pw.cpp)
    - drop the namespace dftu_pw block (moved to dftu_tools_pw.cpp)
    - include "source_pw/module_pwdft/dftu_tools_pw.h" instead of dftu_pw.h
- source/source_pw/module_pwdft/CMakeLists.txt
    - replace dftu_pw.cpp with dftu_tools_pw.cpp + dftu_cal_occ_pw.cpp
- source/Makefile.Objects
    - replace dftu_pw.o with dftu_tools_pw.o + dftu_cal_occ_pw.o
- source/source_pw/module_pwdft/dftu_base.cpp
    - update the pointer comment to reflect new filenames

Effects:
- header/implementation properly paired for the free functions
- dftu_tools_pw.cpp has minimal dependencies, can be linked into
  unit tests without pulling in onsite_proj.h / parallel_reduce.h
- main implementation file (dftu_cal_occ_pw.cpp) retains its git history
  via git mv (rename similarity ~63%, above the 50% threshold)

Verification:
- make module_pwdft incremental build passed (verified by user)
- step 4d proper (rewire tests to call real functions) follows next
