+++ source/source_pw/module_pwdft/CMakeLists.txt
@@ -14,7 +14,8 @@ list(APPEND objects
     op_pw_exx_pot.cpp
     dftu_base.cpp
     dftu_output.cpp
-    dftu_pw.cpp
+    dftu_tools_pw.cpp
+    dftu_cal_occ_pw.cpp
     setup_pot.cpp
     setup_pwrho.cpp
     setup_pwwfc.cpp
